+++ gamevault/Models/AppInfo.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
@@ -44,10 +45,12 @@ namespace gamevault.Models
         ForcedInstallationType,
         SyncSteamShortcuts,
         SyncDiscordPresence,
+        CloudSaves,
         //DevMode
         DevModeEnabled,
-        DevTargetPhalcodeTestBackend
+        DevTargetPhalcodeTestBackend,
         //
+        InstallationId
     }
     public static class AppFilePath
     {
@@ -55,7 +58,7 @@ namespace gamevault.Models
         internal static string OfflineProgress = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/cache/prgs";
         internal static string OfflineCache = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/cache/local";
         internal static string ConfigDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/config";
-        internal static string ThemesLoadDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}\\GameVault\\themes";
+        internal static string ThemesLoadDir = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "GameVault", "themes");
         internal static string WebConfigDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/config/web";
         internal static string UserFile = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/config/user";
         internal static string IgnoreList = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/cache/ignorelist";
@@ -68,7 +71,7 @@ namespace gamevault.Models
             OfflineProgress = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/cache/prgs";
             OfflineCache = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/cache/local";
             ConfigDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/config";
-            ThemesLoadDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}\\GameVault\\debug\\themes";
+            ThemesLoadDir = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "GameVault", "debug", "themes");
             WebConfigDir = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/config/web";
             UserFile = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/config/user";
             IgnoreList = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}/GameVault/debug/cache/ignorelist";