+++ gamevault/PipeServiceHandler.cs
@@ -454,12 +454,16 @@ namespace gamevault
         /// <returns>The game or null if not installed</returns>
         private async Task<Game?> GetInstalledGame(int id)
         {
-            if (!InstallViewModel.Instance.InstalledGames.Any())
+            Game? game = null;
+            await Dispatch(async () =>
             {
-                await MainWindowViewModel.Instance.Library.GetGameInstalls().RestoreInstalledGames();
-            }
-
-            return InstallViewModel.Instance.InstalledGames.Where(g => g.Key.ID == id).Select(g => g.Key).FirstOrDefault();
+                if (!InstallViewModel.Instance.InstalledGames.Any())
+                {
+                    await MainWindowViewModel.Instance.Library.GetGameInstalls().RestoreInstalledGames();
+                }
+                game = InstallViewModel.Instance.InstalledGames.Where(g => g.Key.ID == id).Select(g => g.Key).FirstOrDefault();
+            });
+            return game;
         }
 
         /// <summary>
@@ -513,13 +517,18 @@ namespace gamevault
                 return;
 
             // Ensure we're on the UI thread
-            await Dispatch(() =>
+            await Dispatch(async () =>
             {
+                if (App.Instance.MainWindow != null)
+                {
+                    App.Instance.MainWindow.Activate();
+                    await Task.Delay(500);
+                }
 
                 var gameViewUserControl = new UserControls.GameViewUserControl(game, LoginManager.Instance.IsLoggedIn());
                 // Set the correct UI regardless of if it's visible to let the user manage it
                 MainWindowViewModel.Instance.SetActiveControl(gameViewUserControl);
-                InstallUserControl.PlayGame(game.ID);
+                await InstallUserControl.PlayGame(game.ID);
             });
         }