+++ gamevault/UserControls/GameViewUserControl.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using gamevault.Converter;
 using gamevault.Helper;
+using gamevault.Helper.Integrations;
 using gamevault.Models;
 using gamevault.ViewModels;
 using LiveChartsCore.Measure;
@@ -137,34 +138,6 @@ namespace gamevault.UserControls
             gameID = game.ID;
             this.DataContext = ViewModel;
         }
-        private async void ReloadGameView_Click(object sender, System.Windows.Input.KeyEventArgs e)
-        {
-            if (e.Key != Key.F5)
-                return;
-
-            await ReloadGameView();
-        }
-        private async Task ReloadGameView()
-        {
-            this.IsEnabled = false;
-            try
-            {
-                string result = await WebHelper.GetRequestAsync(@$"{SettingsViewModel.Instance.ServerUrl}/api/games/{gameID}");
-                ViewModel.Game = JsonSerializer.Deserialize<Game>(result);
-                ViewModel.UserProgresses = ViewModel.Game.Progresses.Where(p => p.User.ID != LoginManager.Instance.GetCurrentUser().ID).ToArray();
-                ViewModel.CurrentUserProgress = ViewModel.Game.Progresses.FirstOrDefault(progress => progress.User.ID == LoginManager.Instance.GetCurrentUser()?.ID) ?? new Progress { MinutesPlayed = 0, State = State.UNPLAYED.ToString() };
-            }
-            catch (Exception ex) { }
-            ViewModel.IsInstalled = IsGameInstalled(ViewModel.Game);
-            ViewModel.IsDownloaded = IsGameDownloaded(ViewModel.Game);           
-            PrepareMarkdownElements();
-            this.IsEnabled = true;
-        }
-        public void RefreshGame(Game game)
-        {
-            ViewModel.Game = game;
-            PrepareMarkdownElements();
-        }
         private async void UserControl_Loaded(object sender, RoutedEventArgs e)
         {
             this.Focus();
@@ -184,8 +157,16 @@ namespace gamevault.UserControls
                     catch (Exception ex) { }
                 }
                 ViewModel.IsInstalled = IsGameInstalled(ViewModel.Game);
-                ViewModel.IsDownloaded = IsGameDownloaded(ViewModel.Game);               
+                ViewModel.IsDownloaded = IsGameDownloaded(ViewModel.Game);
                 PrepareMarkdownElements();
+                _ = Task.Run(async () =>
+                {
+                    try
+                    {
+                        ViewModel.CloudSaveMatchTitle = await SaveGameHelper.Instance.SearchForLudusaviGameTitle(ViewModel?.Game?.Metadata?.Title);
+                    }
+                    catch { }
+                });
                 //MediaSlider
                 try
                 {
@@ -201,6 +182,35 @@ namespace gamevault.UserControls
                 uiMediaSlider.Dispose();
             }
         }
+        private async void ReloadGameView_Click(object sender, System.Windows.Input.KeyEventArgs e)
+        {
+            if (e.Key != Key.F5)
+                return;
+
+            await ReloadGameView();
+        }
+        private async Task ReloadGameView()
+        {
+            this.IsEnabled = false;
+            try
+            {
+                string result = await WebHelper.GetRequestAsync(@$"{SettingsViewModel.Instance.ServerUrl}/api/games/{gameID}");
+                ViewModel.Game = JsonSerializer.Deserialize<Game>(result);
+                ViewModel.UserProgresses = ViewModel.Game.Progresses.Where(p => p.User.ID != LoginManager.Instance.GetCurrentUser().ID).ToArray();
+                ViewModel.CurrentUserProgress = ViewModel.Game.Progresses.FirstOrDefault(progress => progress.User.ID == LoginManager.Instance.GetCurrentUser()?.ID) ?? new Progress { MinutesPlayed = 0, State = State.UNPLAYED.ToString() };
+            }
+            catch (Exception ex) { }
+            ViewModel.IsInstalled = IsGameInstalled(ViewModel.Game);
+            ViewModel.IsDownloaded = IsGameDownloaded(ViewModel.Game);
+            PrepareMarkdownElements();
+            this.IsEnabled = true;
+        }
+        public void RefreshGame(Game game)
+        {
+            ViewModel.Game = game;
+            PrepareMarkdownElements();
+        }
+
         private bool IsGameInstalled(Game? game)
         {
             if (game == null)
@@ -228,9 +238,11 @@ namespace gamevault.UserControls
         {
             MainWindowViewModel.Instance.UndoActiveControl();
         }
-        private void GamePlay_Click(object sender, RoutedEventArgs e)
+        private async void GamePlay_Click(object sender, RoutedEventArgs e)
         {
-            InstallUserControl.PlayGame(ViewModel.Game.ID);
+            ((FrameworkElement)sender).IsEnabled = false;
+            await InstallUserControl.PlayGame(ViewModel.Game.ID);
+            ((FrameworkElement)sender).IsEnabled = true;
         }
         private void GameSettings_Click(object sender, RoutedEventArgs e)
         {
@@ -368,6 +380,59 @@ namespace gamevault.UserControls
             }
             catch { }
         }
+        private async void BackupCloudSaves_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                if (!SettingsViewModel.Instance.License.IsActive())
+                {
+                    MainWindowViewModel.Instance.SetActiveControl(MainControl.Settings);
+                    MainWindowViewModel.Instance.Settings.SetTabIndex(4);
+                    return;
+                }
+                MainWindowViewModel.Instance.AppBarText = "Uploading Savegame to the Server...";
+                ((FrameworkElement)sender).IsEnabled = false;
+                bool success = await SaveGameHelper.Instance.BackupSaveGame(ViewModel!.Game!.ID);
+                if (success)
+                {
+                    MainWindowViewModel.Instance.AppBarText = "Successfully synchronized the cloud saves";
+                }
+                else
+                {
+                    MainWindowViewModel.Instance.AppBarText = "Failed to upload your Savegame to the Server";
+                }
+            }
+            catch
+            {
+                MainWindowViewModel.Instance.AppBarText = "Failed to Upload your Savegame to the Server";
+            }
+            ((FrameworkElement)sender).IsEnabled = true;
+        }
+        private async void RestoreCloudSaves_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                if (!SettingsViewModel.Instance.License.IsActive())
+                {
+                    MainWindowViewModel.Instance.SetActiveControl(MainControl.Settings);
+                    MainWindowViewModel.Instance.Settings.SetTabIndex(4);
+                    return;
+                }
+                MainWindowViewModel.Instance.AppBarText = $"Syncing cloud save...";
+                ((FrameworkElement)sender).IsEnabled = false;
+                string installationDir = InstallViewModel.Instance.InstalledGames.First(g => g.Key.ID == ViewModel!.Game!.ID).Value;
+                bool success = await SaveGameHelper.Instance.RestoreBackup(ViewModel!.Game!.ID, installationDir);
+                if (success)
+                {
+                    MainWindowViewModel.Instance.AppBarText = "Successfully synchronized the cloud save";
+                }              
+            }
+            catch
+            {
+                MainWindowViewModel.Instance.AppBarText = "Failed to restore the Savegame";
+            }
+            ((FrameworkElement)sender).IsEnabled = true;
+        }
         #region Markdown        
         private void OpenHyperlink(object sender, System.Windows.Input.ExecutedRoutedEventArgs e)
         {
@@ -399,8 +464,8 @@ namespace gamevault.UserControls
             }
             catch { }
         }
-        #endregion
 
+        #endregion
 
     }
 }